+++ app/src/main/java/com/jgt/wizelinebaz2023/storage/local/room/entities/base/MoviesTable.kt
@@ -17,7 +17,7 @@ data class MoviesTable(
     val homepage:         String? = null,
     val overview:         String? = null,
     val popularity:       Float,
-    val revenue:          Int,
+    val revenue:          Float,
     val runtime:          Int? = null,
     val status:           String,
     val title:            String,
